fix: assign tree's objects by variables,
 moves helper function to utils-tree.js
